Fix dividend yield per year display when split in multiple transactions
The calculation logic works fine for line items like in the transaction
tab, where the individual number of stocks and their specific cost needs
to be taken into consideration.
When calculating the average dividend yield per year, this is done by
summing up the individual rates of return and here the calculation needs
to relate to the total costs, otherwise multiple percentages are summed
up.

Issue: #1879
Issue: https://forum.portfolio-performance.info/t/dividendenrendite-div-jahr-bei-gleichem-wertpapier-in-mehreren-depots/11911
Co-authored-by: inv-trad <59824809+inv-trad@users.noreply.github.com>
Co-authored-by: Andreas Buchen <andreas.buchen@gmail.com>