Avoid destroying punctuation in tooltip wordwrapping
The current word-wrapping code removes trailing punctuation
(except , and .), such as in:

    test test test (test)

where the closing parenthesis is removed in the tooltips,
e.g. for the Note column (securities view).

This is because the last word boundary is after "test",
but afterwards nothing matches.
With the new code, we no longer rely on word boundaries (\b)
and instead require [ \t\n\x0b\r\f,.] as separator.
We add a space (instead of the terminator symbol X) to
ensure that the last line is correctly matched.